requirements.txt
asgiref==3.6.0
attrs==22.2.0
backports.zoneinfo==0.2.1
bcrypt==4.0.1
Bottleneck @ file:///opt/conda/conda-bld/bottleneck_1657175564434/work
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.1
distro==1.8.0
Django==4.1.6
django-rest-framework==0.1.0
djangorestframework==3.14.0
docker==6.0.1
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
idna==3.4
jsonschema==3.2.0
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
mkl-service==2.4.0
mysqlclient==2.1.1
numexpr @ file:///croot/numexpr_1668713893690/work
numpy @ file:///croot/numpy_and_numpy_base_1672336185480/work
packaging==23.0
pandas==1.5.3
paramiko==3.0.0
pycparser==2.21
PyMySQL==1.0.2
PyNaCl==1.5.0
pyrsistent==0.19.3
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-dotenv==0.21.1
pytz @ file:///croot/pytz_1671697431263/work
PyYAML==5.4.1
requests==2.28.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
sqlparse==0.4.3
texttable==1.6.7
urllib3==1.26.14
websocket-client==0.59.0
